move things out of libc-misc that have a better home; make libc-misc work on illumos
